worker/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
yt-dlp==2025.1.15
python-dotenv==1.0.1
